Add DropdownQuestionModel DTO
DropdownQuestionModel, which is a Data Transfer Object (DTO) used for handling Dropdown type questions within the application.

The DropdownQuestionModel includes the following properties:

Choices: Represents a list of string options for the Dropdown question. This property will be used to store the available options for the user to choose from when answering the question.

EnableOtherOption: Indicates whether the "Other" option is enabled for the Dropdown question. If set to true, it allows the user to provide a custom answer in addition to the predefined options in the "Choices" list.

By utilizing the DropdownQuestionModel, the application can effectively handle Dropdown type questions and capture user responses accurately.