Add MultipleChoiceQuestionModel DTO
MultipleChoiceQuestionModel, which is a Data Transfer Object (DTO) used for handling Multiple Choice type questions within the ConsoleProject application.

The MultipleChoiceQuestionModel includes the following properties:

Options: Represents the list of available options for the multiple-choice question. This property will store the various choices that users can select from.

MaximumNumberOfChoicesAllowed: Represents the maximum number of choices allowed for the multiple-choice question. This property helps limit the number of selections a user can make.

EnableOtherOption: Indicates whether an "Other" option is enabled for the multiple-choice question. When enabled, users can enter custom choices that are not in the predefined options list.

The ChosenOptions property is commented out and is not in use currently, as indicated.