Add UpdateQuestionModel DTO
UpdateQuestionModel, which is a Data Transfer Object (DTO) used for updating a question within the ConsoleProject application.

The UpdateQuestionModel includes the following properties:

Response: Represents the user's response to the question. This property will store the user's input or answer to the updated question.

QuestionContent: Represents the content of the question being updated. This property will hold the new content of the question that the user wants to update it to.