Add VideoQuestionModel DTO
 VideoQuestionModel, which is a Data Transfer Object (DTO) used for handling video-related questions within the ConsoleProject application.

The VideoQuestionModel includes the following properties:

AdditionalSubmissionInformation: Represents any additional information the user may want to provide while submitting the video. This property allows users to add comments or notes related to the video submission.

MaxDurationOfVideo: Represents the maximum duration allowed for the video. This property defines the maximum length of the video that users can submit in response to the video-related question.