Add YesOrNoQuestionModel DTO
YesOrNoQuestionModel, which is a Data Transfer Object (DTO) used for handling yes or no questions within the ConsoleProject application.

The YesOrNoQuestionModel includes the following properties:

Choice: Represents the user's choice for the yes or no question. When the user selects 'Yes,' this property will be set to `true`, and when the user selects 'No,' it will be set to `false`.

DisqualifyForNoChoice: Indicates whether selecting 'No' will disqualify the user from further stages. If set to `true`, choosing 'No' will disqualify the user from progressing to the next stage or question.