Add BaseResponse DTO
adds the BaseResponse<T> generic Data Transfer Object (DTO) to the application. 
The BaseResponse<T> class is designed to hold response data with the following properties:

Data: Represents the generic data returned in the response. It can hold any type of data.

Status: Indicates the status of the response, where 'true' typically signifies success and 'false' indicates failure.

Message: Provides an optional message or description related to the response, which can be used for informative purposes.