Add QuestionType enumeration
adds the QuestionType enumeration to the application. The enum represents different types of questions that can be used in the application's survey or questionnaire functionality. The enum contains various question types, such as Paragraph, ShortAnswer, YesOrNo, Dropdown, MultipleChoice, Date, Number, VideoQuestion, and FileUpload. These types will be used to categorize and differentiate the various types of questions that can be presented to users during the survey or questionnaire process.