Add Program and other related classes
adds the Program class and related classes to the application. The Program class represents a program with additional information, such as ProgramTitle, Description, ApplicantSkills, Benefits, ApplicationCriteria, and more. The ProgramLocation class is used to store the location details of a program. The AdditionalInformation class contains common properties related to a program, including ProgramType, ProgramStart, ApplicationStart, ApplicationEnds, ApplicantQualification, ProgramLocations, DurationInMonths, and MaxApplications. These classes will be used to manage program data and applications in the application.