Add BaseModel and ProfileBase class
Add BaseModel and ProfileBase abstract classes to the application. The BaseModel class serves as the base model for all entities and includes an Id property marked with the [Key] attribute to represent a unique identifier for each entity. The ProfileBase class acts as the base model for profile-related entities and contains properties like Mandatory and Hidden, which indicate whether specific profile attributes are mandatory and should be hidden, respectively. These base classes provide common functionality and properties that can be inherited and used in various entity and profile-related classes within the application.