Add VideoInterviewStage and VideoInterviewQuestion models
Add VideoInterviewStage and VideoInterviewQuestion, to the application. The VideoInterviewStage class holds a list of VideoInterviewQuestion objects, representing a stage with multiple video interview questions. On the other hand, the VideoInterviewQuestion class holds information about a single video interview question, including the text of the question, additional submission information, maximum video duration, and the deadline to submit the response in the number of days. These models will be used to manage video interview-related data within the application.