Add FileUploadQuestion model
Add the FileUploadQuestion model to the application. The FileUploadQuestion class represents a question that allows users to upload files. The model includes a property named FilePath, which is a string that will hold the file path where the uploaded file will be stored. This model will be used to manage file upload questions within the application.