Add MultipleChoiceQuestion model
Add the MultipleChoiceQuestion model to the application. The MultipleChoiceQuestion class represents a question that allows users to select multiple choices from a predefined list of options. The model includes properties for 'Options', which is a list of available choices, 'MaximumChoicesAllowed', an integer specifying the maximum number of choices allowed to be selected, 'EnableOtherOption', a boolean flag to enable the "Other" option, and 'ChosenOptions', a list that holds the user's selected choices. This model will be used to manage multiple-choice questions within the application.