Add Application Context
Adds the ApplicationContext class to the Web API Project Context class. The class inherits from DbContext and manages the database context for the application. It includes DbSet properties for the Program, Question, Stage, and Application entities, and it also includes the OnModelCreating method to configure the entity models and seed initial data for the Stage entity.